refactor(src/ui/types): rename some files
