+++ packages/ui/src/ui/store/selectors/job/statistics.ts
@@ -1,5 +1,3 @@
-import {JobStatistic} from '../../../types/job';
 import {RootState} from '../../reducers';
 
-export const getRawStatistic = (state: RootState) =>
-    state.job.general.job.statistics as JobStatistic;
+export const getRawStatistic = (state: RootState) => state.job.general.job.statistics;
